@task 90843: EM Jenkins plug-in to execute test scenario jobs

+++ environment-manager-client/src/main/java/com/parasoft/em/client/api/Systems.java
@@ -1,18 +1,10 @@
 /*
- * $RCSfile$
- * $Revision$
- *
- * Comments:
- *
  * (C) Copyright ParaSoft Corporation 2013.  All rights reserved.
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF ParaSoft
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
- *
- * $Author$          $Locker$
- * $Date$
- * $Log$
  */
+
 package com.parasoft.em.client.api;
 
 import java.io.IOException;
@@ -23,5 +15,5 @@ import net.sf.json.JSONObject;
  */
 public interface Systems {
     JSONObject getSystems() throws IOException;
-    JSONObject getSystem(int id) throws IOException;
+    JSONObject getSystem(long id) throws IOException;
 }
