+++ environment-manager-client/src/main/java/com/parasoft/em/client/impl/SystemsImpl.java
@@ -1,18 +1,10 @@
 /*
- * $RCSfile$
- * $Revision$
- *
- * Comments:
- *
  * (C) Copyright ParaSoft Corporation 2013.  All rights reserved.
  * THIS IS UNPUBLISHED PROPRIETARY SOURCE CODE OF ParaSoft
  * The copyright notice above does not evidence any
  * actual or intended publication of such source code.
- *
- * $Author$          $Locker$
- * $Date$
- * $Log$
  */
+
 package com.parasoft.em.client.impl;
 
 import java.io.IOException;
@@ -31,7 +23,7 @@ public class SystemsImpl extends JSONClient implements Systems {
         return doGet("api/v1/systems", true);
     }
 
-    public JSONObject getSystem(int id) throws IOException {
+    public JSONObject getSystem(long id) throws IOException {
         return doGet("api/v1/systems/" + id);
     }
